data_generator/requirements.txt
pandas
numpy
python-dotenv
sqlalchemy
psycopg2-binary
tqdm
kafka-python